factor out framework dependent code in iterators
